+++ src/samples/randoms/mod.rs
@@ -19,13 +19,13 @@ use crate::id_targeted::IdTargeted;
 /// Sum     calculates the sum of the contained numbers
 #[derive(Debug, PartialEq, Eq)]
 pub struct Randoms {
-    pub id: u64,
+    pub id: usize,
     pub numbers: Vec<u64>,
 }
 
 impl Randoms {
-    pub fn new(id: u64) -> Self {
-        let mut rng = Xoshiro256Plus::seed_from_u64(id);
+    pub fn new(id: usize) -> Self {
+        let mut rng = Xoshiro256Plus::seed_from_u64(id as u64);
         let numbers = (0..10000).map(|_| rng.next_u64()).collect();
         Self { id, numbers }
     }
@@ -48,7 +48,7 @@ impl Randoms {
 }
 
 impl IdTargeted for Randoms {
-    fn id(&self) -> u64 {
+    fn id(&self) -> usize {
         self.id
     }
 }
