+++ src/samples/randoms/randoms_api/mean_request.rs
@@ -3,11 +3,11 @@ use crate::id_targeted::IdTargeted;
 /// This is the message sent to request that the Randoms struct (with the given id) calculates the mean of the random numbers it contains.
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct MeanRequest {
-    pub id: u64,
+    pub id: usize,
 }
 
 impl IdTargeted for MeanRequest {
-    fn id(&self) -> u64 {
+    fn id(&self) -> usize {
         self.id
     }
 }
